requirements.txt
requests==2.29.0
beautifulsoup4
streamlit
numpy
Pillow
keras~=2.6.0
tensorflow~=2.15.0